Lots changing currently broken
Basic profile page has been created. the intake code has been started for this sheet. Table has not been made yet. getUser.php has been reworked to rely on methods. This is a better method but index.php and other need to be changed to work with the new format. Will be better over all in the end